Properly save selection state when the selection root is the editor
In cases where the cursor is at the beginning/end of the text or when all of the text is selected, window.getSelection() can specify the selection with respect to the editor instead of specific text nodes.
This change adds support for saving selection state in save() when this is the case.
Fixes #75