fixing self closing characters logic
When a typed character is an open bracket or quote and there is a selection, it wraps the selection in brackets or quotes, regardless of characters before/after the selection.
The close bracket is generated when an open bracket isn't escaped and there is any closing bracket (`)]}`) or a white-space character (` \t\n`) before the open bracket or at the end.
The close quote is generated only if there is no any non-white-space character before/after the typed quote.
The code `.replace(/\r\n?/g, '\n')` isn't related to this issue, it just a correction. It's a professional way to normalize lines ending to `\n`.